requirements/dev.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
flake8==4.0.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.5
openapi-codec==1.3.2
pandas==1.3.5
pathspec==0.9.0
pip==21.3.1
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-decouple==3.5
python-environ==0.4.54
pytz==2021.3
regex==2021.11.10
requests==2.26.0
setuptools==58.5.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
